[exporter/prometheusremotewrite] allocate smaller buffers for prometheus remote write batch time series (#34271)
**Description:** 
Adjusting buffer allocations in
prometheusremotewriteexporter.batchTimeSeries

Rather than allocating the maximum possible size buffer each time we
memoize the last buffer size that we actually used and allocate a 2x
sized buffer for the next request. This ensures that we tend to allocate
buffers sized appropriately for the requests and still avoid array
resizing in most cases (unless the data point size changes by more than
2x from one request to the next).

This reduces memory allocations and memory usage significantly in our
testing when you have large batches.

In the case that the request sizes never meet the max request size
configured for the exporter the existing behavior will remain and we
will continue to allocate a single buffer with a size equal to the
number of time series / metric metadata objects passed in. There will
still be a benefit of not allocating a huge requests buffer in this case
since instead of allocating a buffer with a size equal to the number of
total objects passed in we default to a buffer of size 10 or 2x the last
number of requests whichever is larger.

**Link to tracking Issue:** 
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/34269
**Testing:**
We have validated this fix in our test environment and existing unit
tests cover this code. The observable behavior is unchanged as this just
tweaks the pre-allocated size of the buffers. I can add some unit tests
to validate the memoized buffer sizes once we agree on the proper place
to store that memoization.

**Documentation:**
No documentation added as this is a perf optimization that doesn't have
any user facing impact beyond reducing memory usage.