[chore][receiver/nginx] Update README for consistency (#34923)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
This just cleans up the README a bit:
1. Make case of `NGINX` consistent. Happy to change it to `nginx` if
that's preferred.
2. Remove empty `Details` section
3. Cleanup description of the `collection_interval` config option. This
is the same config option many other receivers have, so I don't think
the concept needs explained in-depth here.
4. Remove disclaimer that the component's status is beta and
configuration is subject to change. The README's header states this
component is beta, so I don't think it needs to be repeated.