[chore] Force require.EventuallyWithT to fail properly (#35032)
I noticed on #34720 and #35026 that execution of the test continued
beyond a failure of `require.EventuallyWithT`. Based on the description
alone, I would expect that using `assert` within
`require.EventuallyWithT` should cause execution to stop if the
assertion fails, but it appears this may not be the case. However, this
change apparently works as intended.