[chore] [exporter/clickhouse] Upgrade trace stability to beta (#35186)
**Description:**

This PR upgrades the status of traces to _beta_ to properly reflect its
stability. This is similar to the recent upgrade PR for logs (#33615)

Configuration options are unlikely to change for traces, and the
exporter has proven to be stable for both traces and logs. Table schema
may be subject to change as query optimizations are added in ClickHouse.
Metrics will remain in alpha for now.

Thanks!

**Documentation:**
- Updated metadata.yaml
- Updated README
- Re-ran `make generate`