[sumconnector] Metrics to metrics and logs to metrics (tests / test data) (#35434)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
Adds tests for metrics to metrics and logs to metrics
follows this pr:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/34797

This is a lot of lines but it is mostly repetition and uses the same
standard structure as the previously merged trace to metric tests.

I dont think this needs any changelog but let me know if it does for
some reason! 😸

**Link to tracking Issue:** #32669 

**Testing:** <Describe what testing was performed and which tests were
added.>
Adds testing for metrics to metrics and logs to metrics using same
methodology as previous PR

**Documentation:** No changes to docs. This is just tests.

---------

Co-authored-by: Curtis Robert <crobert@splunk.com>