[chore] Fix `make update-otel` on macOS (#35587)
**Description:**

When working on #35580, it seemed that `make update-otel` was not
properly updating the `builder-config.yaml` files. I traced this down to
the `updatehelper` script not working as intended. This turned out to be
because it uses some features of the `sed` utility which do not exist on
macOS, causing no updates to be made.

The `\s` (whitespace) regex class is a GNU extension for `sed`; the
POSIX-compatible equivalent is `[[:space:]]` ([related SO
question](https://stackoverflow.com/questions/18840175/find-and-replace-with-spaces-using-sed-mac-terminal)).

Moreover, on macOS, the `-i` (in-place) option requires an argument,
even if empty; otherwise the `-e` following it is parsed as that
argument, creating an unnecessary backup file ([related SO
question](https://stackoverflow.com/questions/4247068/sed-command-with-i-option-failing-on-mac-but-works-on-linux)).

**Testing:**

I manually tested this change as part of developing the aforementioned
PR; the script seems to work on macOS now. It would be good for someone
to test this change on Linux to make sure nothing has broken there.