[chore][pkg/stanza] refactor: introduce emit.Token struct (#36260)
This makes the code clearer by encapsulating the token's body and
attributes in a single structure. It should make future change clearer
when the emit callback will be changed to accept a collection of tokens
as opposed to a single token.

The Sink type could use some refactoring as well, but I'm not doing it
here to keep the changes to the minimum for clarity and ease of code
review.