parser/keyvalue: handle escaped quotes when parsing (#36176)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
Currently, the keyvalue parser does handle escaped quotes and instead
will parse this as independent '\', '"' characters. This results in
unexpected breakages between fields for strings like:
> <key>="...\\" ..."
Here, the backslash will be appended to the result pair, while the (now
un-)escaped quotation will result in the pair being terminated early.

Add handling of escaped quotation marks (for both ", ') in the keyvalue
parser along with a testcase to exercise this functionality.

---------

Signed-off-by: Ethan Adams <eadams@netflix.com>
Co-authored-by: Daniel Jaglowski <jaglows3@gmail.com>