[pkg/ottl] Change context inferrer to use functions and enums as hints (#36869)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This PR is part of
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/29017,
and a spin-off from
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/36820.
It changes the existing context inferrer logic to also take into
consideration the functions and enums used on the statements.
(https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/36820#issuecomment-2542028806)

New logic:
- Find all `path`, function names(`editor`, `converter`), and
`enumSymbol` on the statements
- Pick the highest priority context (same existing logic) based on the
`path.Context` values
- If the chosen context does not support all used functions and enums,
it goes through it's lower contexts (wide scope contexts that does
support the chosen context as a path context) testing them and choosing
the first one that supports them.
- If no context that can handle the functions and enums is found, the
inference fail and an empty value is returned.


The parser collection was adapted to support the new context inferrer
configuration requirements.

**Other important changes:**

Currently, it's possible to have paths to contexts root objects, for
example: `set(attributes["body"], resource)`. Given `resource` has no
dot separators on the path, the grammar extracts it into the
`path.Fields` slice, letting the `path.Context` value empty. Why? This
grammar behaviour is still necessary to keep backward compatibility with
paths without context, otherwise it would start requiring contexts for
all paths independently of the parser configuration.

Previous PRs didn't take this edge case into consideration, and a few
places needed to be changed to address it:
 - Context inferrer (`getContextCandidate`)
 - Parser `prependContextToStatementPaths` function.
- Reusable OTTL contexts (`contexts/internal`) (not part of this PR, it
will be fixed by
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/36820)

When/If we reach the point to deprecate paths _without_ context, all
those conditions can be removed, and the grammar changed to require and
extract the `path` context properly.
 
<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue

https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/29017

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Unit tests

<!--Describe the documentation added.-->
#### Documentation
No changes

<!--Please delete paragraphs that you did not use before submitting.-->