[chore] Add scoped GH action to test affected components on Windows  (#37106)
#### Description
Many times changes that break Windows CI are merged since it is not
apparent that they will break on Windows. Given the cost of running
Windows CI for every PR the label `Run Windows` is only added in few
cases. This change adds make targets that allow to run make commands
only for tests and dependencies affected by files being changed.

It uses a simple detection of go files that were changed and look to the
import commands to find the components that depend directly on the
package being changed.

#### Link to tracking issue
Relates to #36788

#### Testing
Validated on my fork, e.g.:
https://github.com/pjanotti/opentelemetry-service-contrib/actions/runs/12659871185

#### Documentation
N/A