[chore] 3 Improvements to scoped-tests action (#37145)
#### Description

1. Completely skip the `scoped-tests` if there are no changes to go
files (avoids checking out and setting the repo and not running any
tests). Test on my fork:
https://github.com/pjanotti/opentelemetry-service-contrib/actions/runs/12698095207/job/35395815043?pr=31
2. Make the step to build `gotestsum` work on non-Windows - this needs
further investigation, since the `test` target depends on `gotestsum`,
but, as seem yesterday it was failing to build the tool (at least on
Windows GH runner, it didn't repro on my box). Test on my fork:
https://github.com/pjanotti/opentelemetry-service-contrib/actions/runs/12697644120/job/35394429408
3. On `Makefile.Common` account for the case when an internal package is
not used by any other package.