[receiver/gitlab] add tracing via webhook skeleton (#36838)
#### Description
This PR adds the structure and trace skeleton for a new and already
accepted Gitlabreceiver. (thanks @atoulme for sponsoring this!)

The Gitlabreceiver aligns very closely with the Githubreceiver and this
PR mostly mirrors the change from this PR:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/36632

I'm working together with @adrielp on building out the Gitlabreceiver.
More PRs to introduce metrics and actual tracing functionality are about
to follow with subsequent PRs.

#### Link to tracking issue

https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/35207

#### Testing
Added basic tests and built the component to test that the health check
endpoint, when tracing is enabled, operates correctly.

#### Documentation
Docs how to configure the Gitlabreceiver via webhooks have been added.
While the Gitlabreceiver can be configured after this PR, it will not
actually do anything since it is under development and just the skeleton
PR.