[chore] fix failing syslog load test (#37197)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

The loadtest is failing due to missing year in the timestamp. The
timestamp `Oct 11 22:14:15` was until Oct 10 2024 calculated with the
year 2023 (`1697062455000000000`) and since then, it by default takes
the year 2024, so it results in the following timestamp
`1728684855000000000` and causes the test to fail.

Failing run on main:
https://github.com/open-telemetry/opentelemetry-collector-contrib/actions/runs/12757989830/job/35563296082

## Note

failing lint test will be resolved in a separate PR
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/37194

Signed-off-by: odubajDT <ondrej.dubaj@dynatrace.com>