[receiver/googlecloudpubsub] Add support for encoding extensions (#37109) (#37137)
#### Description
Added support for encoding extensions. Setting the encoding field in the
config now references the extension. If it didn't find the extension, it
will fall back to searching the internal encoders.

To make the build in encoders consistent with the extensions, they now
have the same interface.

#### Link to tracking issue
Fixes #37109

#### Testing
- Added tests for the receiver to handle the new extensions and the
build in encoders
- Removed the tests in the configuration, as they are now handled in the
receiver tests

#### Documentation
Reworked the encoding section of the README, with an example of a
text_encoding extension