[chore] Update otel version (#37808)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
This is a PR to verify the change in #37706: After running `make
update-otel`, `builder-config.yaml` files are correctly updated in this
PR.

Feel free to close this PR if it's too large to review. (I have no
permission to run the `update-otel` GitHub action and have to run it
manually).
cc @bogdandrutu @codeboten @mx-psi @songy23 
<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Relevant to #37703 

<!--Describe what testing was performed and which tests were added.-->
#### Testing
n/a

<!--Describe the documentation added.-->
#### Documentation
n/a

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Signed-off-by: Chao <chao.weng@appdynamics.com>
Co-authored-by: Pablo Baeyens <pablo.baeyens@datadoghq.com>