[tailsamplingprocessor] Support external decision cache implementations (#37035)
#### Description

Adding a feature. This PR adds support for external implementations of
the decision cache. This allows the collector (or another service using
the processor) to supply an alternative decision cache based on
alternative algorithms or external services like memcached without
needing to explicitly add support for all possible options in the
processor.

It re-uses the existing function option pattern and only exposes two
options for now: `WithSampledDecisionCache` and
`WithNonSampledDecisionCache`. I've avoided exporting other options to
avoid bloating the external interface without a concrete use case. The
majority of changes are cleanup from the refactoring to move `Option`
values into the `Config` struct instead of in a variadic parameter on
`newTracesProcessor`.

---------

Co-authored-by: Yuna Verheyden <yuna.verheyden@posteo.net>