[receiver/prometheus] Add fallback_scrape_protocol during config validation (#38018)
#### Description
During the release process, we noticed a breaking change in the
Prometheus receiver caused by
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/36873.

In that PR, I tried adding a fallback scrape protocol by default
everytime the PrometheusReceiver was built, but it turned out that
config validation happened even before the component was created. And
the collector fails startup with invalid configuration

This PR moves the addition of scrape protocol to the validation step.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #37902

<!--Describe what testing was performed and which tests were added.-->
#### Testing

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->