[chore] move evaluation of file lists to tasks to reduce Makefile setup (#37951)
This change is brought about by the realization that with #37945 no code
is directly under the root module. This has the unexpected impact that
some of the Makefile set up started to show warnings.

It turns out that we do quite a bit of reading files and go packages in
the Makefile. This has a performance impact. We also have incorrect code
in there, such as the computation of `ALL_PKG_DIRS` only returns
`github.com/open-telemetry/opentelemetry-collector-contrib/cmd/checkapi`.