[extension/sumologicextension] Sanitize version information (#37689)
#### Description

The format of versions reported by Sumo Logic's distribution of the
OpenTelemetry collector recently changed. The new format is not
currently supported by the sumologicextension in this repository. This
PR adds a new function to sanitize the version before it is used to send
data to the Sumo Logic backend service.

#### Testing

Added unit tests to test variations of possible collector version
output. Used in the last two releases of Sumo Logic's distribution of
the OpenTelemetry Collector.

---------

Signed-off-by: Justin Kolberg <amd.prophet@gmail.com>
Co-authored-by: Francis Guimond <francis@sensu.io>
Co-authored-by: Raj Nishtala <rnishtala@sumologic.com>
Co-authored-by: Raj Nishtala <113392743+rnishtala-sumo@users.noreply.github.com>