Remove high-cardinality attribute from AWS Container Insight metrics (#37697)
#### Description
Per code inspection it looks like `tags` was being used as a convenience
feature to pass the timestamp when converting the metrics to OTLP.
However, timestamp should not be a resource attribute due to causing
high-cardinality time series.

This change keeps the current usage of `tags`, but, ensures that
timestamp is not added as a resource attribute. Code owners should
consider if later the timestamp should be passed outside the `tags` map
- a change much larger than the current one.

#### Link to tracking issue
Fixes #35861

#### Testing
Updated respective tests.

#### Documentation
Changelog added.