[chore] Remove usages of extensionauthtest.MockClient (#38401)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

All usages of `extensionauthtest.MockClient` can be replaced by a simple
`extensionauth.NewClient` + passing the appropriate options.

There is one use case that is maybe useful to have a helper for: errors,
which are used in `confighttp` tests.

This PR removes usages of `extensionauthtest.MockClient` so that we can
remove the `MockClient` and replace it with something that allows for a
sealed interface/struct for `extensionauth.Client`

Co-authored-by: Antoine Toulme <atoulme@splunk.com>