[redisstorageextension] allow redis key prefix to be specified in config (#37677)
#### Description
The purpose of this PR is to allow the user to specify the redis key
prefix. The lack of this feature has become problematic when using
multiple replicas in a kubernetes cluster - until now, the key prefix
has been based on the storage extension and receiver names, which are
the same for each replica because they share the same configuration and
therefore overwrite the same key in redis. With the changes in this PR,
the user can specify the prefix as an environmental variable (which can
have different value for each replica), for example.

#### Testing
opentelemetry-collector-contrib (0.118.0) was built with this change and
tested on a kubernetes cluster

Config used for testing:
```
redis_storage:
  endpoint: redis.monitoring.svc.cluster.local:6379
  prefix: ${K8S_NODE_NAME}
```

#### Documentation
Updated README.md to add new config parameter.

---------

Co-authored-by: Antoine Toulme <atoulme@splunk.com>