[receiver/prometheus] feat: add Prometheus API server (#33875)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
This is a new PR instead of this one:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/32646
which has more relevant discussion.
This PR adds the Prometheus API server as part of the Prometheus
receiver if enabled through the receiver config.
It supports calling the `/config`, `/targets`, `/targets/metadata`,
`/scrape_pools`, and `/metrics` paths for more debugging information
about the underlying Prometheus scrape manager and discovery manager.

**Link to tracking Issue:** <Issue number if applicable>
https://github.com/open-telemetry/wg-prometheus/issues/63

**Testing:** 
Added tests for the prometheus receiver configuration, for each path
that the API supports, and for the server shutting down correctly. Also
tested out with running the opentelemetry-collector agent with the
prometheus receiver.


**Documentation:** 
Added to the prometheus receiver README about how to configure the
setting.

---------

Co-authored-by: David Ashpole <dashpole@google.com>