googlecloudexporter: add ability to write to custom MRs (#38331)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This change adds a featuregate that allows writing to custom monitored
resources based on presence of the `gcp.resource_type` resource
attribute.

The default mapping from OTel resources to MRs are used as a fallback.

To enable it, use the `exporter.googlecloud.CustomMonitoredResources`
featuregate.
<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #38102

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Added unit tests

<!--Describe the documentation added.-->
#### Documentation
N/A

<!--Please delete paragraphs that you did not use before submitting.-->

Signed-off-by: Ridwan Sharif <ridwanmsharif@google.com>