Initial commit for setting up a new component: k8slog receiver (#24439)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
k8slog receiver provides full functionality to collect logs in k8s
containers. It is an all-in-one solution, which supports to:

- collect logs from files inside k8s containers via daemonset
- support docker and cri-containerd
- support many docker graph drivers / snapshotters
- collect logs from stdout of k8s containers via daemonset
- filter containers by metadata
- extract metadata of containers
- collect file logs in k8s containers via sidecar

If we want to collect logs from k8s, we can use this component.

**Link to tracking Issue:** <Issue number if applicable>

#23339

**Testing:** <Describe what testing was performed and which tests were
added.>

Tests for unmarshaling configuration have been added.

**Documentation:** <Describe the documentation added.>

Docs for configurations and overall design has been added.

---------

Co-authored-by: Chris Mark <chrismarkou92@gmail.com>