[reciever/prometheusremotewritereceiver] Handle multiple timeseries with same metric name+type+unit (#38453)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This PR belongs to part of the Linux foundation mentee program. Here we
are dealing with data points slices that must be added to the same
metric when the incoming time series have some attributes in common.
Like same resource, metricName, scopeName, scopeVersion, unitRef and
timeseries type. The reference to this implementation can be found
https://opentelemetry.io/docs/specs/otel/metrics/data-model/#opentelemetry-protocol-data-model

Besides that. We are creating 2 new test cases. The first one, to
validate the described behavior above.

The second is to validate an error case. When the unitRef passed doesn't
match with the symbols slice, causing a panic. This test case is
important to guarantee that the error is being handled correctly.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes part of #37277. Bullet point `Handle multiple timeseries with same
metric name+type+unit`

---------

Co-authored-by: Arthur Silva Sens <arthursens2005@gmail.com>