[chore] internal/kafka: introduce kafkatest package (#38882)
#### Description

This package will hold test-related code for Kafka components.

Initially, the package provides a single function for creating a fake
Kafka cluster using franz-go's kfake, along with a
configkafka.ClientConfig for connecting to it. We use this in new tests
for NewSaramaClient.

#### Link to tracking issue

N/A

#### Testing

N/A, only adding test code.

#### Documentation

N/A