[chore] [exporter/elasticsearch] Allow storing state in the OTel encoder (#38732)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

For
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/38606,
we want to be able to store state within the otel encoder.
We can't do that, as we encoder is root methods, not within a struct.

So this PR changes that behavior to keep encoders in memory, so they can
introduce shared state such as cache.