[chore] issue 33461 (#38884)
#### Description

The JSON DATA section is used to add the CHANGE_FROM_PREVIOUS_WEEK for
each voice of the report.

The code was removed because the generate JSON would get the issue
content past the allowed limit.

With these changes we are not adding the JSON data if it will bring the
issue chart past the allowed limit.

#### Link to tracking issue

Fixes: #33461 

#### Testing

Currently testing from my local machine with:

```
 act -j get_issues -e <(echo '{"repository": {"owner": {"login": "lucalanziani"}}}') -s GITHUB_TOKEN=<my token>
 ```
 
The result can be seen here: https://github.com/LucaLanziani/opentelemetry-collector-contrib/issues/17