[cmd/opampsupervisor] Control Collectors with only the OpAMP extension (#38809)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Right now three components are required for the Supervisor to control a
Collector:
* nopreceiver
* nopexporter
* opampextension

The `service.AllowNoPipelines` feature gate allows the Collector to be
started without pipelines, which was released in v0.122.0. This can be
used to start the Collector with only the OpAMP extension during the
bootstrapping process, obviating the need for the nopreceiver and
nopexporter.

---------

Co-authored-by: Evan Bradley <evan-bradley@users.noreply.github.com>