[receiver/azuremonitorreceiver] feat: add subscription name resource attribute (#39029)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
The subscription name is often more speaking for people playing with
metrics, than the subscription id which is an uuid.
Hence this PR. 

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Linked with #38895. I prefer to have a separate PR for the add of
subscription name in resource attributes to have one PR by feature.

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Added unit tests + new mock API to mock the new get by subscription
calls.

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->

Signed-off-by: Célian Garcia <celian.garcia@amadeus.com>
Co-authored-by: Antoine Toulme <atoulme@splunk.com>