[receiver/azuremonitorreceiver] feat: option to use azmetrics getBatch API (#38895)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
Let me introduce the Batch Metrics API
https://learn.microsoft.com/en-us/rest/api/monitor/metrics-batch/batch?view=rest-monitor-2023-10-01&tabs=HTTP

The advantages of using ``azmetrics`` batch metrics API over the
``armmonitor`` metrics API are described in details in the issue + in
the README.md

Note that this is a revival of a previous PR that we made last year, and
that has not been followed up properly but we're using that forked code
in the Amadeus company for months with good results.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Relates #38651

<!--Describe what testing was performed and which tests were added.-->
#### Testing
✅  Manual tests + mocked unit tests

<!--Describe the documentation added.-->
#### Documentation

✅  Added config field documentation in README.md
<!--Please delete paragraphs that you did not use before submitting.-->

Signed-off-by: Célian Garcia <celian.garcia@amadeus.com>